add: Search reservations by date range

+++ src/utils/dateFormatHelper.js
@@ -13,7 +13,7 @@ export function formateDateToLocale(date) {
   // remove the Z from date
   const formattedDate = new Date(date.slice(0, date.length - 1));
   return new Intl.DateTimeFormat(language, {
-    weekday: "long",
+    weekday: "short",
     year: "numeric",
     month: "short",
     day: "numeric",
